requirements.txt
livekit-agents[cartesia,elevenlabs,turn-detector]~=1.0
livekit==0.17.1
livekit-agents==0.9.1
livekit-api==0.7.0
livekit-plugins-deepgram==0.6.7
livekit-plugins-openai==0.8.5
livekit-plugins-rag==0.2.2
livekit-plugins-silero==0.6.4
livekit-plugins-noise-cancellation~=0.2
python-dotenv
llama-index-core~=0.11.10
llama-index-vector-stores-pinecone~=0.2.0
llama-index-embeddings-openai~=0.2.0
llama-index-llms-openai
pinecone-client~=5.0.0